+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## TBD
+
+### Enhancements
+
+- Update the version of the BugSnag CLI we're using to 2.6.1 [35](https://github.com/bugsnag/bugsnag-gradle-plugin/pull/35)
+
 ## 0.1.0 (2024-03-27)
 
 Initial Release of BugSnag Gradle Plugin