+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Enhancements
 
-- Update the version of the BugSnag CLI we're using to 2.6.1 [35](https://github.com/bugsnag/bugsnag-gradle-plugin/pull/35)
+- Update bugsnag-cli dependency to [v2.6.1](https://github.com/bugsnag/bugsnag-cli/blob/next/CHANGELOG.md#261-2024-09-18) [35](https://github.com/bugsnag/bugsnag-gradle-plugin/pull/35)
 
 ## 0.1.0 (2024-03-27)