+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0 (2025-07-14)
+
+### Enhancements
+
+- Update bugsnag-cli dependency to [v3.2.0](https://github.com/bugsnag/bugsnag-cli/blob/next/CHANGELOG.md#320---2025-07-02) [#51](https://github.com/bugsnag/bugsnag-gradle-plugin/pull/51)
+
 ## 0.4.0 (2025-04-15)
 
 ### Enhancements